Sh4 recompiler base added ;) Recompiles around half cpu , and no fpu Constant Elimination is performed , nothing more
Crazy Taxi runs , but hangs after sometime ingame...
Bios runs , but has some bad textures...
Everything else iv tried runs allright(mailny kos demos ect)

TODO : 
Fix up the emiiter Lib/Use another one
Finish up the rest opcodes
Add FPU 


git-svn-id: file:///home/skmp/Desktop/public_svn/nullDC@34 d5fdb02f-ce0d-0410-91b1-fe443107cc8c

+++ trunk/nulldc/nullDC/dc/sh4/sh4_fpu.cpp
@@ -448,8 +448,8 @@ sh4op(i1111_nnn0_1111_1101)
 	 
 	if (fpscr.PR==0)
 	{
-	fr[n | 0] = sin(real_pi);
-	fr[n | 1] = cos(real_pi);
+	fr[n | 0] = (float)sin(real_pi);
+	fr[n | 1] = (float)cos(real_pi);
 	
 	CHECK_FPU_32(fr[n]);
 	CHECK_FPU_32(fr[n+1]);
